kana-writing/frontend/requirements.txt
streamlit==1.28.2
pillow==10.1.0
streamlit-drawable-canvas==0.9.3
requests==2.31.0
